backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.8.0
pandas==2.2.2
openpyxl==3.1.5
tabula-py==2.9.3
PyMuPDF==1.24.0
Pillow==10.4.0
pydantic==2.9.0
python-dotenv==1.0.1
aiofiles==24.1.0
sqlalchemy
slowapi
plotly
bcrypt
requests
